Util: create two algorithm to change systematicly start and end-time of releases
i created two static methods. each public-methods represent one algorithm
the fist algorithm cutting of the waste iteration. for example. the relase
has a start time on 2012/3/1 and end time 2012/4/1. so we have two iteration.
the first start with release-start and end on 2012/3/15. the second iteration
start after first iteration's end and end on release end.
so we change the release start from 2012/3/1 to 2012/3/16. so we cut of the
first iteration, because it is not include in the ne release

the second algorithm works with the durations and percentage of total duration
the unchanged release has a duration of 31 days. the new release only 15 days.
so we calculate the percentage of all iterations and slide it on the timeline

Change-Id: Ib5aaa84965bc75d0977f18509db888768dd0f32f